1/backup/BackendAPI/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
SQLAlchemy>=2.0.23
alembic>=1.12.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.0
httpx>=0.24.1
reportlab>=4.0.4
Pillow>=10.0.0
qrcode[pil]>=7.4.2
# Advanced features dependencies
redis>=5.0.0
aioredis>=2.0.0
prometheus-client>=0.18.0
structlog>=23.1.0
pendulum>=2.1.2
numpy>=1.26.0
pandas>=2.2.0
backoff>=2.2.0
semver>=3.0.0
psutil>=7.0.0
docker>=7.1.0
psycopg2-binary>=2.9.10
# Threat detection dependencies
dnspython>=2.4.0
user-agents>=2.2.0
